First cut implementation: Arrays and Strings
Completed parsing arrays and strings. First cut implementation.
Started with a `RedisProtocolParser` struct and implemented the parsing
methods inside `impl` block of that. That looks more correct. Previously
we were using the parsing methods as `scoped` methods in the `impl`
block of the `Value`. That was not so semantically correct.